% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{add.metadata}
\alias{add.metadata}
\title{Add metadata}
\usage{
add.metadata(genomeset_results, genome_metadata, genome_metadata_idcol)
}
\arguments{
\item{genomeset_results}{genomeset_results}

\item{genome_metadata}{genome_metadata}

\item{genome_metadata_idcol}{genome_metadata_idcol}
}
\value{
results
}
\description{
Add metadata
}
