% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{assert.rules}
\alias{assert.rules}
\title{Apply rules to a set of detected genes}
\usage{
assert.rules(genes)
}
\arguments{
\item{genes}{}
}
\value{

}
\description{
Apply rules to a set of detected genes
}
