% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{assert.rulesforgenome}
\alias{assert.rulesforgenome}
\title{Apply rules to a set of detected genes}
\usage{
assert.rulesforgenome(genes, domains, cov_threshold = 80)
}
\arguments{
\item{genes}{domains cov_threshold}
}
\value{

}
\description{
Apply rules to a set of detected genes
}
