% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_conservation.R
\name{build.taxa.model}
\alias{build.taxa.model}
\title{Build lme model}
\usage{
build.taxa.model(trait_matrix, trait, species = TRUE, optimizer = "nlm")
}
\value{
model
}
\description{
Build lme model
}
