% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{calc_intergenome_variance}
\alias{calc_intergenome_variance}
\title{Calculate interguild variance}
\usage{
calc_intergenome_variance(
  genomeset_results,
  trait_granularity = "trait_matrixatgranularity3",
  method = "wishart",
  normby = "genome_length",
  binarytype = "logical",
  clustering_method = "ward.D",
  outdir,
  ncores = 1
)
}
\arguments{
\item{genomeset_results}{}
}
\description{
Calculate interguild variance
}
