% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{calc_mixeddist}
\alias{calc_mixeddist}
\title{Calculate distances between genomes.}
\usage{
calc_mixeddist(
  trait_matrix,
  idcol = "id",
  col2ignore = c("genome_length", "mingentime"),
  method = "wishart",
  binarytype = "logical",
  byrow = 1,
  verbose = TRUE
)
}
\arguments{
\item{trait_matrix_bin}{}
}
\description{
Calculate distances between genomes.
}
