% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{cluster_traitmatrix}
\alias{cluster_traitmatrix}
\title{Genome by trait matrix heatmap.}
\usage{
cluster_traitmatrix(
  trait_matrix,
  idcol = "id",
  annot_cols = c("mingentime", "ogt"),
  granularity = 3,
  cutree_rows = NA,
  cutree_cols = NA,
  clustering_distance_rows,
  clustering_distance_cols,
  width = width,
  height = height,
  heatmap_width,
  heatmap_height,
  row_dend_width,
  column_dend_height,
  rightannotation_width,
  topannotation_height,
  bottomannotation_height,
  outdir,
  dataset,
  pdf = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{trait_matrix}{}
}
\description{
Genome by trait matrix heatmap.
}
