% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_utils.R
\name{convert_traitdatatype}
\alias{convert_traitdatatype}
\title{Make sure binary traits are of type logical}
\usage{
convert_traitdatatype(genomeset_results, binarytype = "logical")
}
\arguments{
\item{genomeset_results}{genomeset_results}
}
\value{
genomeset_results
}
\description{
Make sure binary traits are of type logical
}
