% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{count.traitsforgenome}
\alias{count.traitsforgenome}
\title{Apply rules to a set of detected genes}
\usage{
count.traitsforgenome(rules_asserted, trait)
}
\arguments{
\item{genes_detected}{}
}
\value{

}
\description{
Apply rules to a set of detected genes
}
