% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{define_guilds}
\alias{define_guilds}
\title{Define guild matrix.}
\usage{
define_guilds(
  trait_matrix,
  hclust_rows,
  hclust_cols,
  idcol = "id",
  annot_cols = c("mingentime", "ogt"),
  nguilds,
  guildsizecutoff = 50,
  traitvariability_threshold = 0.01,
  width,
  height,
  heatmap_width,
  heatmap_height,
  rightannotation_width,
  topannotation_height,
  bottomannotation_height,
  outdir,
  dataset,
  verbose = TRUE
)
}
\arguments{
\item{trait_matrix}{trait_matrix}

\item{nguilds}{nguilds}

\item{clusters_traitmatrix}{clusters_traitmatrix}
}
\description{
Define guild matrix.
}
