% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{detect.genes.fromdomtblout}
\alias{detect.genes.fromdomtblout}
\title{Apply rules to a set of detected genes}
\usage{
detect.genes.fromdomtblout(domtblout, cov_threshold = 80)
}
\arguments{
\item{domtblout}{cov_threshold}
}
\value{

}
\description{
Apply rules to a set of detected genes
}
