% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmmpackage.R
\name{download.arcbacribosomal}
\alias{download.arcbacribosomal}
\title{Prepare database for PFAM models for ribosomal proteins}
\usage{
download.arcbacribosomal(arcbacribosomalhmmdb_file)
}
\description{
Prepare database for PFAM models for ribosomal proteins
}
