% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmmpackage.R
\name{download.dbcan}
\alias{download.dbcan}
\title{Prepare dbcan database (download and subselect)}
\usage{
download.dbcan(dbcan_version = 8, dbcanhmmdb_selectids_file, dbcanhmmdb_file)
}
\description{
Prepare dbcan database (download and subselect)
}
