% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmmpackage.R
\name{download.microtrait}
\alias{download.microtrait}
\title{Prepare microtrait database (download and subselect)}
\usage{
download.microtrait(microtraithmmdb_file)
}
\description{
Prepare microtrait database (download and subselect)
}
