% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{extract.traits.parallel}
\alias{extract.traits.parallel}
\title{Run microtrait for genomes in parallel using mclapply}
\usage{
\method{extract}{traits.parallel}(
  fna_files,
  out_dirs = dirname(fna_files),
  ncores = floor(parallel::detectCores() * 0.7),
  save_tempfiles = F,
  type = "genomic"
)
}
\arguments{
\item{fna_files}{cache_dir out_dir}
}
\value{
results
}
\description{
Run microtrait for genomes in parallel using mclapply
}
