% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{fetch.results}
\alias{fetch.results}
\title{Fetch results for a single genome}
\usage{
fetch.results(rds_file, id, type)
}
\arguments{
\item{rds_file}{id type}
}
\value{

}
\description{
Fetch results for a single genome
}
