% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{filter.bycoverage.domtblout}
\alias{filter.bycoverage.domtblout}
\title{Read a file created through the '----domtblout' option}
\usage{
filter.bycoverage.domtblout(domtblout, cov_by, cov_threshold = 60)
}
\arguments{
\item{file}{Filename}
}
\value{
data.frame
}
\description{
Read a file created through the '----domtblout' option
}
