% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{gp.resultsummary}
\alias{gp.resultsummary}
\title{Fetch growthpred results for a single genome}
\usage{
gp.resultsummary(file)
}
\arguments{
\item{file}{file}
}
\value{

}
\description{
Fetch growthpred results for a single genome
}
