% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmms_fromrules}
\alias{hmms_fromrules}
\title{microtraithmm_fromrules.}
\format{
A data frame with 13 variables:
\describe{
\item{microtraithmm-name}{}
\item{microtraithmm-dbxref_kegg}{}
\item{microtraithmm-dbxref_ec}{}
\item{microtraithmm-dbxref_TC}{}
\item{microtraithmm-description}{}
\item{microtraithmm-notes}{}
\item{score}{}
\item{fscore}{}
\item{tpr}{}
\item{tnr}{}
\item{fpr}{}
\item{fnr}{}
\item{accuracy}{}
}
}
\usage{
hmms_fromrules
}
\description{
HMMs (Hidden Markov Models) underlying microTrait pipeline.
This data provides information on the database cross references for
each HMM and the performance of the HMM using KEGG orthologs as a
benchmark.
}
\keyword{datasets}
