% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{make.genomeset.results}
\alias{make.genomeset.results}
\title{Combine microtrait outputs for multiple genomes into trait matrices (genomes x traits)}
\usage{
make.genomeset.results(
  rds_files,
  ids = NULL,
  growthrate = T,
  optimumT = T,
  ncores = 1
)
}
\arguments{
\item{rds_files}{rds_files}

\item{ncores}{ncores}
}
\value{
genomeset_results
}
\description{
Combine microtrait outputs for multiple genomes into trait matrices (genomes x traits)
}
