% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_traits.R
\name{map.traits.fromdomtblout}
\alias{map.traits.fromdomtblout}
\title{Apply rules to a set of detected genes}
\usage{
map.traits.fromdomtblout(microtrait_domtblout, dbcan_domtblout)
}
\arguments{
\item{domtblout_file1}{domtblout_file2 out_file}
}
\value{
mapped traits from hmm hit tables.
import dplyr
}
\description{
Apply rules to a set of detected genes
}
