% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rRNA.R
\name{nuclcheck}
\alias{nuclcheck}
\title{base = "/Users/ukaraoz/Work/microtrait/code/github/test/microtrait/inst/extdata/genomic/Desulfotomaculum_kuznetsovii_dsm_6115_ASM21470v1_dna_toplevel"
x <- readFASTA(file.path(base, "translated.faa"))
x = x\link{1:3}
protcheck(x) # TRUE
protcheck(paste(x, "Z", sep = "")) # FALSE}
\usage{
nuclcheck(x)
}
\description{
base = "/Users/ukaraoz/Work/microtrait/code/github/test/microtrait/inst/extdata/genomic/Desulfotomaculum_kuznetsovii_dsm_6115_ASM21470v1_dna_toplevel"
x <- readFASTA(file.path(base, "translated.faa"))
x = x\link{1:3}
protcheck(x) # TRUE
protcheck(paste(x, "Z", sep = "")) # FALSE
}
