% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmmpackage.R
\name{prep.hmmmodels}
\alias{prep.hmmmodels}
\title{Prepare microtrait database (download and subselect)}
\usage{
prep.hmmmodels(output_dir = system.file("extdata", package = "microtrait"))
}
\description{
Prepare microtrait database (download and subselect)
}
