% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{read.domtblout}
\alias{read.domtblout}
\title{Read a file created through the '----domtblout' option}
\usage{
read.domtblout(file)
}
\arguments{
\item{file}{(Required). hmmscan output file to be read.}
}
\value{
data frame of class tibble
}
\description{
Read a file created through the '----domtblout' option
}
