% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{rule2parts}
\alias{rule2parts}
\title{Return genes that are part of a rule.}
\usage{
rule2parts(rule_boolean)
}
\arguments{
\item{rule}{(Required). microtrait rule.}
}
\value{
rule parts
}
\description{
Return genes that are part of a rule.
}
