% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{rulename2ruleboolean}
\alias{rulename2ruleboolean}
\title{Get genes from rule}
\usage{
rulename2ruleboolean(rule)
}
\arguments{
\item{rule}{(Required). microtrait rule.}
}
\value{

}
\description{
Get genes from rule
}
