% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rules}
\alias{rules}
\title{microtraitrule_table.}
\format{
A data frame with 13 variables:
\describe{
\item{microtraithmm-name}{}
\item{microtraithmm-rule}{}
\item{microtraithmm-name-quoted}{}
\item{microtraithmm-ruleunwrapped}{}
}
}
\usage{
rules
}
\description{
Rules underlying microTrait pipeline.
Each rule is a logical rule using presence/absence of microTrait HMMs.
}
\keyword{datasets}
