% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_conservation.R
\name{run.consentrait.singletrait}
\alias{run.consentrait.singletrait}
\title{Run consentrait script}
\usage{
run.consentrait.singletrait(
  trait,
  traitmatrix,
  tree,
  treeroot,
  nbootstrap = 100,
  nproc = 40,
  perc.share.cutoff = 90,
  outdir,
  consentraitbin = "/global/homes/u/ukaraoz/bin/consentrait.R",
  run = FALSE
)
}
\value{
result
}
\description{
Run consentrait script
}
