% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{run.parallel}
\alias{run.parallel}
\title{Run microtrait for genomes in parallel using mclapply}
\usage{
run.parallel(fna_files, out_dirs, save_tempfiles = F, type = "genomic", ncores)
}
\arguments{
\item{fna_files}{cache_dir out_dir}
}
\value{
results
}
\description{
Run microtrait for genomes in parallel using mclapply
}
