% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mingentime.R
\name{run.predictGrowth}
\alias{run.predictGrowth}
\title{predictGrowth from codon usage bias}
\usage{
run.predictGrowth(
  cds_file,
  proteins_file,
  mode = "full",
  temperature = "none",
  training_set = "vs",
  depth_of_coverage = NULL,
  fragments = FALSE
)
}
\arguments{
\item{cds_file}{proteins_file}
}
\value{

}
\description{
predictGrowth from codon usage bias
}
