% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{set.geneprofile}
\alias{set.geneprofile}
\title{Set gene profiles using detected genes}
\usage{
set.geneprofile(genes)
}
\arguments{
\item{genes}{gene list}
}
\value{

}
\description{
Set gene profiles using detected genes
}
