% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{splitSeqsToFiles}
\alias{splitSeqsToFiles}
\title{Split DNA sequences in a fasta file into single sequence files.}
\usage{
splitSeqsToFiles(
 
    fasta_file = "/Users/ukaraoz/Work/microtrait/code/inst/extdata/GIG_1C_0-10cm_R1-3.slf.fa",
  suffix = "temp",
  outDir = tempdir(),
  ncores = 4
)
}
\arguments{
\item{fasta_file}{FASTA filename.}

\item{suffix}{a word to add to each file created. Default is "temp".}

\item{outDir}{directory to write the output file. Default is tempdir()}

\item{ncores}{number of cores to use}
}
\value{
a vector of filename names created
}
\description{
Split DNA sequences in a fasta file into single sequence files.
}
