% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_utils.R
\name{trait.continuous2binary}
\alias{trait.continuous2binary}
\title{Convert continuous traits to binary}
\usage{
trait.continuous2binary(feature_matrix)
}
\value{
results
}
\description{
Convert continuous traits to binary
}
