% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_utils.R
\name{trait.normalize}
\alias{trait.normalize}
\title{Normalize count traits}
\usage{
trait.normalize(genomeset_results, normby = "Estimated Size")
}
\arguments{
\item{genomeset_results}{}
}
\value{
genomeset_results
}
\description{
Normalize count traits
}
