% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{trait2trait_corr}
\alias{trait2trait_corr}
\title{Analyze trait2trait correlations.}
\usage{
trait2trait_corr(
  trait_matrix_bin,
  idcol = "id",
  annot_cols = c("mingentime", "ogt"),
  verbose = T,
  outdir,
  dataset
)
}
\arguments{
\item{trait_matrix_bin}{}
}
\description{
Analyze trait2trait correlations.
}
