% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_utils.R
\name{unwrap.rules}
\alias{unwrap.rules}
\title{Unwraps complex rules.}
\usage{
unwrap.rules(rule_table)
}
\arguments{
\item{rule}{(Required). microtrait rule.}
}
\value{
rule parts
}
\description{
Unwraps complex rules.
}
