% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{write.genomeset.results}
\alias{write.genomeset.results}
\title{Add metadata}
\usage{
write.genomeset.results(genomeset_results, out_dir, datasetid)
}
\arguments{
\item{rds_dir}{rds_dir}

\item{ncores}{ncores}
}
\value{
genomeset_results
}
\description{
Add metadata
}
