% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{add.metadata}
\alias{add.metadata}
\title{Add metadata}
\usage{
add.metadata(genomeset_results, gp_results, genome_metadata)
}
\arguments{
\item{genomeset_results}{genomeset_results}

\item{gp_results}{gp_results}

\item{genome_metadata}{genome_metadata}
}
\value{
results
}
\description{
Add metadata
}
