% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{combine.results}
\alias{combine.results}
\title{Combine results for multiple genomes}
\usage{
combine.results(rds_files, ids = NULL, type, ncores = 1)
}
\arguments{
\item{rds_files}{type}
}
\value{

}
\description{
Combine results for multiple genomes
}
