% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitanalysis_utils.R
\name{compute.prevalence}
\alias{compute.prevalence}
\title{Count prevalence of traits, rules, or hmms}
\usage{
\method{compute}{prevalence}(feature_matrix, type, idcol = "id")
}
\value{
results
}
\description{
Count prevalence of traits, rules, or hmms
}
