% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtrait.R
\name{make.genomeset.results}
\alias{make.genomeset.results}
\title{Add metadata}
\usage{
make.genomeset.results(rds_files, ids = NULL, ncores)
}
\arguments{
\item{ncores}{ncores}

\item{rds_dir}{rds_dir}
}
\value{
genomeset_results
}
\description{
Add metadata
}
