% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_hmmpackage.R
\name{prep.hmmmodels}
\alias{prep.hmmmodels}
\title{Prepare microtrait database}
\usage{
prep.hmmmodels(output_dir = system.file("extdata", package = "microtrait"))
}
\description{
\code{prep.hmmmodels} downloads and prepares hmm models required by microtrait.
}
\details{

}
